backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Blockchain
blockcypher==1.0.93
requests==2.31.0
bitcoinlib==0.6.14

# Security & Crypto
pydantic==2.5.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.0

# Background Tasks
apscheduler==3.10.4

# API & HTTP
httpx==0.26.0
aiohttp==3.9.1

# Utilities
python-multipart==0.0.6
email-validator==2.1.0

# CORS
fastapi-cors==0.0.6

# Logging
python-json-logger==2.0.7

# Testing (optional)
pytest==7.4.4
pytest-asyncio==0.23.3
